requirements.txt
selenium==4.37.0
pytest==8.4.2
allure-pytest==2.15.0
Faker==37.12.0